separated component and added loading indicator
